use a u32 instead of a u64 as Attributes storage
A macro is used to generate the mapping from attributes
to SGR values and avoid duplicating the attribute names
in code.

+++ src/style/attributes.rs
@@ -3,7 +3,7 @@ use std::ops::{BitAnd, BitOr, BitXor};
 
 /// a bitset for all possible attributes
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-pub struct Attributes(u64);
+pub struct Attributes(u32);
 
 impl From<Attribute> for Attributes {
     fn from(attribute: Attribute) -> Self {
